Renaming, verified that all 3 flavors build
