+++ Samples/CustomEditControl/cpp-winui/App.idl
@@ -1,6 +1,6 @@
 ﻿// Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-namespace CustomEditControlWinAppSDK
+namespace CustomEditControlCpp
 {
 }
