+++ Samples/CustomEditControl/cpp-winui/App.xaml
@@ -2,10 +2,10 @@
  Licensed under the MIT License. -->
 
 <Application
-    x:Class="CustomEditControlWinAppSDK.App"
+    x:Class="CustomEditControlCpp.App"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:local="using:CustomEditControlWinAppSDK">
+    xmlns:local="using:CustomEditControlCpp">
     <Application.Resources>
         <ResourceDictionary>
             <ResourceDictionary.MergedDictionaries>
