+++ Samples/CustomEditControl/cpp-winui/App.xaml.cpp
@@ -11,8 +11,8 @@ using namespace Windows::Foundation;
 using namespace Microsoft::UI::Xaml;
 using namespace Microsoft::UI::Xaml::Controls;
 using namespace Microsoft::UI::Xaml::Navigation;
-using namespace CustomEditControlWinAppSDK;
-using namespace CustomEditControlWinAppSDK::implementation;
+using namespace CustomEditControlCpp;
+using namespace CustomEditControlCpp::implementation;
 
 // To learn more about WinUI, the WinUI project structure,
 // and more about our project templates, see: http://aka.ms/winui-project-info.
