+++ Samples/CustomEditControl/cpp-winui/CustomEditControlCpp.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.32002.261
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "CustomEditControlWinAppSDK", "CustomEditControlWinAppSDK.vcxproj", "{4F349AF3-4A2D-439C-BB5E-5E899BD234CA}"
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "CustomEditControlCpp", "CustomEditControlCpp.vcxproj", "{4F349AF3-4A2D-439C-BB5E-5E899BD234CA}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
