+++ Samples/CustomEditControl/cpp-winui/MainWindow.xaml
@@ -2,11 +2,10 @@
  Licensed under the MIT License. -->
 
 <Window
-    x:Name="CustomEditControlWindow"
-    x:Class="CustomEditControlWinAppSDK.MainWindow"
+    x:Class="CustomEditControlCpp.MainWindow"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:local="using:CustomEditControlWinAppSDK"
+    xmlns:local="using:CustomEditControlCpp"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
     mc:Ignorable="d">
