+++ Samples/CustomEditControl/cpp-winui/Package.appxmanifest
@@ -12,7 +12,7 @@
     Version="1.0.0.0" />
 
   <Properties>
-    <DisplayName>CustomEditControlWinAppSDK</DisplayName>
+    <DisplayName>CustomEditControlCpp</DisplayName>
     <PublisherDisplayName>tyduck</PublisherDisplayName>
     <Logo>Assets\StoreLogo.png</Logo>
   </Properties>
@@ -31,8 +31,8 @@
       Executable="$targetnametoken$.exe"
       EntryPoint="$targetentrypoint$">
       <uap:VisualElements
-        DisplayName="CustomEditControlWinAppSDK"
-        Description="CustomEditControlWinAppSDK"
+        DisplayName="CustomEditControlCpp"
+        Description="CustomEditControlCpp"
         BackgroundColor="transparent"
         Square150x150Logo="Assets\Square150x150Logo.png"
         Square44x44Logo="Assets\Square44x44Logo.png">
