+++ Samples/CustomEditControl/cs-winui/CustomEditControlCs/App.xaml
@@ -2,10 +2,10 @@
  Licensed under the MIT License. -->
 
 <Application
-    x:Class="CustomEditControlSampleCs.App"
+    x:Class="CustomEditControlCs.App"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:local="using:CustomEditControlSampleCs">
+    xmlns:local="using:CustomEditControlCs">
     <Application.Resources>
         <ResourceDictionary>
             <ResourceDictionary.MergedDictionaries>
