+++ Samples/CustomEditControl/cs-winui/CustomEditControlCs/App.xaml.cs
@@ -6,7 +6,7 @@ using Microsoft.UI.Xaml;
 // To learn more about WinUI, the WinUI project structure,
 // and more about our project templates, see: http://aka.ms/winui-project-info.
 
-namespace CustomEditControlSampleCs
+namespace CustomEditControlCs
 {
     /// <summary>
     /// Provides application-specific behavior to supplement the default Application class.
