+++ Samples/CustomEditControl/cs-winui/CustomEditControlCs/Assets/LockScreenLogo.scale-200.png
