+++ Samples/CustomEditControl/cs-winui/CustomEditControlCs/Assets/SplashScreen.scale-200.png
