+++ Samples/CustomEditControl/cs-winui/CustomEditControlCs/Assets/Square150x150Logo.scale-200.png
