+++ Samples/CustomEditControl/cs-winui/CustomEditControlCs/Assets/Square44x44Logo.scale-200.png
