+++ Samples/CustomEditControl/cs-winui/CustomEditControlCs/Assets/StoreLogo.png
