+++ Samples/CustomEditControl/cs-winui/CustomEditControlCs/Assets/Wide310x150Logo.scale-200.png
