+++ Samples/CustomEditControl/cs-winui/CustomEditControlCs/Package.appxmanifest
@@ -12,7 +12,7 @@
     Version="1.0.0.0" />
 
   <Properties>
-    <DisplayName>CustomEditControlSampleCs</DisplayName>
+    <DisplayName>CustomEditControlCs</DisplayName>
     <PublisherDisplayName>tyduck</PublisherDisplayName>
     <Logo>Assets\StoreLogo.png</Logo>
   </Properties>
@@ -31,8 +31,8 @@
       Executable="$targetnametoken$.exe"
       EntryPoint="$targetentrypoint$">
       <uap:VisualElements
-        DisplayName="CustomEditControlSampleCs"
-        Description="CustomEditControlSampleCs"
+        DisplayName="CustomEditControlCs"
+        Description="CustomEditControlCs"
         BackgroundColor="transparent"
         Square150x150Logo="Assets\Square150x150Logo.png"
         Square44x44Logo="Assets\Square44x44Logo.png">
