+++ Samples/CustomEditControl/cs-winui/CustomEditControlCs/app.manifest
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <assembly manifestVersion="1.0" xmlns="urn:schemas-microsoft-com:asm.v1">
-  <assemblyIdentity version="1.0.0.0" name="CustomEditControlSampleCs.app"/>
+  <assemblyIdentity version="1.0.0.0" name="CustomEditControlCs.app"/>
 
   <application xmlns="urn:schemas-microsoft-com:asm.v3">
     <windowsSettings>
